Change so current account isn't unlinked if the discord account is already registered

+++ src/utils/newRegistration.ts
@@ -15,7 +15,7 @@ export function newRegistration(dcclient, d2client, dccode, d2code, res){
             let id = x.membership_id;
             if (d2client.DB.map(e => e.bungieId).includes(id)) {
                 d2client.DB.forEach(user => {
-                    if (user.bungieId === id) {
+                    if (user.bungieId === id && dcuser.id !== user.discordUser.id) {
                         d2client.DB.delete(id)
                         removeAccountRoles(user.discordUser.id, dcclient, d2client)
                     }
